First working version

+++ src/args.rs
@@ -4,10 +4,10 @@ use clap::Parser;
 #[command(version, long_about = None)]
 pub struct Args {
     #[arg(short = 'H', long, default_value_t = 10)]
-    pub height: u32,
+    pub height: usize,
 
     #[arg(short = 'W', long, default_value_t = 10)]
-    pub width: u32,
+    pub width: usize,
 
     #[arg(short = 'S', long, default_value_t = Args::get_default_seed())]
     pub seed: String,
